rename board/position to board/validation
